+++ server/src/app.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
-import GungiController from './gateway/controllers/gungi.controller';
-import GungiUsecaseModule from './usecases/gungi-usecase.module';
-import { DataServicesModule } from './frameworks/data-services/data-service-module';
+import GungiController from './gateway/controllers/Gungi.controller';
+import GungiUsecaseModule from './usecases/GungiUsecase.module';
+import { DataServicesModule } from './frameworks/data-services/DataService.module';
 
 @Module({
   imports: [
