+++ server/src/domain/Gungi.ts
@@ -5,9 +5,9 @@ import GomaOki from './GomaOki';
 import SIDE from './constant/SIDE';
 import GOMA from './constant/GOMA';
 import Coordinate from './Coordinate';
-import { Event, SurrenderEvent } from './events/event';
+import { Event, SurrenderEvent } from './events/Event';
 import DeadArea from './DeadArea';
-import { GungiData } from '../frameworks/data-services/gungi-data';
+import { GungiData } from '../frameworks/data-services/GungiData';
 
 class Gungi {
   private _senteGomaOki: GomaOki;
