+++ server/src/frameworks/data-services/DataService.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { GungiRepositoryModule } from './gungiRepository.module';
+import { GungiRepositoryModule } from './GungiRepository.module';
 
 @Module({
   imports: [GungiRepositoryModule],
