+++ server/src/frameworks/data-services/dao/Dao.ts
@@ -1,0 +1,5 @@
+export interface Dao<T> {
+  findById(id: string): Promise<T>;
+
+  save(data: T): Promise<void>;
+}
