--- server/src/frameworks/data-services/dao/dao.ts
@@ -1,4 +1,0 @@
-export abstract class Dao<T> {
-  abstract findById(id: string): Promise<T>;
-  abstract save(data: T): Promise<void>;
-}
