+++ server/src/frameworks/view-model/GungiViewModel.ts
@@ -1,4 +1,4 @@
-import { Event, SurrenderEvent } from '../../domain/events/event';
+import { Event, SurrenderEvent } from '../../domain/events/Event';
 
 function surrender(events: Event[]) {
   const target = events[0] as SurrenderEvent;
