+++ server/src/usecases/Presenter.ts
@@ -1,4 +1,4 @@
-import { Event } from '../domain/events/event';
+import { Event } from '../domain/events/Event';
 
 export default interface Presenter {
   present(events: Event[]): any;
