+++ server/test/e2e/e2e.e2e-spec.ts
@@ -3,10 +3,10 @@ import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
 import LEVEL from '../../src/domain/constant/LEVEL';
 import { AppModule } from '../../src/app.module';
-import { DataServicesModule } from '../../src/frameworks/data-services/data-service-module';
-import { GungiRepository } from '../../src/frameworks/data-services/gungi-repository';
-import GungiDataModel from '../../src/frameworks/data-services/data-model/gungi-data-model';
-import { GungiData } from '../../src/frameworks/data-services/gungi-data';
+import { DataServicesModule } from '../../src/frameworks/data-services/DataService.module';
+import { GungiRepository } from '../../src/frameworks/data-services/GungiRepository';
+import GungiDataModel from '../../src/frameworks/data-services/data-model/GungiDataModel';
+import { GungiData } from '../../src/frameworks/data-services/GungiData';
 import { randomUUID } from 'crypto';
 import SIDE from '../../src/domain/constant/SIDE';
 import { Db, MongoClient } from 'mongodb';
